requirements.txt
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
et_xmlfile==2.0.0
getmac==0.9.5
idna==3.11
invoke==2.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==8.1.0
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
paramiko==4.0.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.19.1
scp==0.15.0
six==1.17.0
speedtest-cli==2.1.3
textfsm==2.1.0
urllib3==2.6.3
